+++ VirtualList/VirtualList.js
@@ -165,6 +165,30 @@ VirtualList.propTypes = /** @lends sandstone/VirtualList.VirtualList.prototype *
 	'data-spotlight-id': PropTypes.string,
 
 	/**
+	 * Disable voice control feature of component.
+	 *
+	 * @type {Boolean}
+	 * @public
+	 */
+	'data-webos-voice-disabled': PropTypes.bool,
+
+	/**
+	 * Activates the component for voice control.
+	 *
+	 * @type {Boolean}
+	 * @public
+	 */
+	'data-webos-voice-focused': PropTypes.bool,
+
+	/**
+	 * The voice control group label.
+	 *
+	 * @type {String}
+	 * @public
+	 */
+	'data-webos-voice-group-label': PropTypes.string,
+
+	/**
 	 * The layout direction of the list.
 	 *
 	 * @type {('horizontal'|'vertical')}
@@ -609,6 +633,30 @@ VirtualGridList.propTypes = /** @lends sandstone/VirtualList.VirtualGridList.pro
 	 * @private
 	 */
 	'data-spotlight-id': PropTypes.string,
+
+	/**
+	 * Disable voice control feature of component.
+	 *
+	 * @type {Boolean}
+	 * @public
+	 */
+	'data-webos-voice-disabled': PropTypes.bool,
+
+	/**
+	 * Activates the component for voice control.
+	 *
+	 * @type {Boolean}
+	 * @public
+	 */
+	'data-webos-voice-focused': PropTypes.bool,
+
+	/**
+	 * The voice control group label.
+	 *
+	 * @type {String}
+	 * @public
+	 */
+	'data-webos-voice-group-label': PropTypes.string,
 
 	/**
 	 * The layout direction of the list.