+++ internal/Picker/Picker.js
@@ -1,5 +1,5 @@
 import classnames from 'classnames';
-import {forward, stop, stopImmediate} from '@enact/core/handle';
+import {forward, forwardCustom, stop, stopImmediate} from '@enact/core/handle';
 import {is} from '@enact/core/keymap';
 import platform from '@enact/core/platform';
 import {cap, clamp, Job, mergeClassNameMaps} from '@enact/core/util';
@@ -60,7 +60,6 @@ const selectDecIcon = selectIcon('decrementIcon', 'triangledown', 'triangleleft'
 const forwardBlur = forward('onBlur'),
 	forwardFocus = forward('onFocus'),
 	forwardKeyDown = forward('onKeyDown'),
-	forwardMouseDown = forward('onMouseDown'),
 	forwardKeyUp = forward('onKeyUp'),
 	forwardWheel = forward('onWheel');
 
@@ -146,6 +145,22 @@ const PickerBase = class extends ReactComponent {
 		 * @public
 		 */
 		'aria-valuetext': PropTypes.string,
+
+		/**
+		 * Determines which key to adjust the picker's value for the joined horizontal one.
+		 *
+		 *  * `'enter'` allows the user to use the enter key to adjust the picker's value
+		 *  * `'arrow'` allows the user to use the left or right keys to adjust the picker's value.
+		 *
+		 * The default value for joined horizontal picker is `'enter'`.
+		 * If [orientation]{@link sandstone/internal/Picker.PickerBase#orientation} is `'vertical'` or
+		 * [joined]{@link sandstone/internal/Picker.PickerBase#joined} is undefined or is `false`, this prop is ignored.
+		 *
+		 * @type {('enter'|'arrow')}
+		 * @default 'enter'
+		 * @public
+		 */
+		changedBy: PropTypes.oneOf(['enter', 'arrow']),
 
 		/**
 		 * Children from which to pick
@@ -246,8 +261,10 @@ const PickerBase = class extends ReactComponent {
 
 		/**
 		 * Determines the user interaction of the control. A joined picker allows the user to use
-		 * the arrow keys to adjust the picker's value. The user may no longer use those arrow keys
-		 * to navigate, while this control is focused. A split control allows full navigation,
+		 * the arrow keys or the enter key to adjust the picker's value.
+		 * It depends on [changedBy]{@link sandstone/internal/Picker.PickerBase.changedBy}
+		 * whether to use the arrow keys or the enter key.
+		 * A split control allows full navigation,
 		 * but requires individual ENTER presses on the incrementer and decrementer buttons.
 		 * Pointer interaction is the same for both formats.
 		 *
@@ -417,6 +434,7 @@ const PickerBase = class extends ReactComponent {
 
 	static defaultProps = {
 		accessibilityHint: '',
+		changedBy: 'enter',
 		orientation: 'horizontal',
 		spotlightDisabled: false,
 		step: 1,
@@ -471,6 +489,7 @@ const PickerBase = class extends ReactComponent {
 
 	computeNextValue = (delta) => {
 		const {
+			changedBy,
 			joined,
 			min,
 			max,
@@ -478,7 +497,7 @@ const PickerBase = class extends ReactComponent {
 			value,
 			wrap
 		} = this.props;
-		const horizontalJoined = orientation === 'horizontal' && joined;
+		const horizontalJoined = orientation === 'horizontal' && joined && changedBy === 'enter';
 		const shouldWrap = horizontalJoined || wrap;
 
 		return shouldWrap ? wrapRange(min, max, value + delta) : clamp(min, max, value + delta);
@@ -552,14 +571,6 @@ const PickerBase = class extends ReactComponent {
 
 	emulateMouseUp = new Job(this.clearPressedState, 175);
 
-	handleMouseDown = (ev) => {
-		const {joined, orientation} = this.props;
-		forwardMouseDown(ev, this.props);
-
-		if (joined && orientation === 'horizontal') {
-			this.setIncPickerButtonPressed();
-		}
-	};
 
 	handleUp = () => {
 		if (this.props.joined && (this.pickerButtonPressed !== 0 || this.state.pressed !== 0)) {
@@ -568,12 +579,16 @@ const PickerBase = class extends ReactComponent {
 	};
 
 	handleDown = () => {
-		const {joined, orientation} = this.props;
+		const {changedBy, joined, orientation} = this.props;
+
+		if (joined && orientation === 'horizontal' && changedBy === 'enter') {
+			this.setIncPickerButtonPressed();
+		}
 
 		if (joined && this.pickerButtonPressed === 1) {
 			this.handleIncrement();
 
-			if (orientation === 'vertical') {
+			if (orientation === 'vertical' || changedBy === 'arrow') {
 				this.emulateMouseUp.start();
 			}
 		} else if (joined && this.pickerButtonPressed === -1) {
@@ -634,6 +649,7 @@ const PickerBase = class extends ReactComponent {
 
 	handleKeyDown = (ev) => {
 		const {
+			changedBy,
 			joined,
 			onSpotlightDown,
 			onSpotlightLeft,
@@ -649,30 +665,34 @@ const PickerBase = class extends ReactComponent {
 
 			const directions = {
 				up: this.setIncPickerButtonPressed,
-				down: this.setDecPickerButtonPressed
+				down: this.setDecPickerButtonPressed,
+				right: this.setIncPickerButtonPressed,
+				left: this.setDecPickerButtonPressed
 			};
 
 			const isVertical = orientation === 'vertical' && (isUp(keyCode) || isDown(keyCode));
-			const isHorizontal = orientation === 'horizontal' && isEnter(keyCode);
+			const isHorizontal = orientation === 'horizontal' && changedBy === 'enter' && isEnter(keyCode);
+			const isHorizontalArrow = orientation === 'horizontal' && changedBy === 'arrow' && (isRight(keyCode) || isLeft(keyCode));
 
-			if (isVertical) {
+			if (isVertical || isHorizontalArrow) {
 				directions[direction]();
 			} else if (isHorizontal) {
 				this.setIncPickerButtonPressed();
 			} else if (orientation === 'horizontal' && isDown(keyCode) && onSpotlightDown) {
-				onSpotlightDown(ev);
+				forwardCustom('onSpotlightDown')(null, this.props);
 			} else if (orientation === 'horizontal' && isUp(keyCode) && onSpotlightUp) {
-				onSpotlightUp(ev);
+				forwardCustom('onSpotlightUp')(null, this.props);
 			} else if (orientation === 'vertical' && isLeft(keyCode) && onSpotlightLeft) {
-				onSpotlightLeft(ev);
+				forwardCustom('onSpotlightLeft')(null, this.props);
 			} else if (orientation === 'vertical' && isRight(keyCode) && onSpotlightRight) {
-				onSpotlightRight(ev);
+				forwardCustom('onSpotlightRight')(null, this.props);
 			}
 		}
 	};
 
 	handleKeyUp = (ev) => {
 		const {
+			changedBy,
 			joined,
 			orientation
 		} = this.props;
@@ -682,8 +702,9 @@ const PickerBase = class extends ReactComponent {
 		if (joined && !this.props.disabled) {
 			const isVertical = orientation === 'vertical' && (isUp(keyCode) || isDown(keyCode));
 			const isHorizontal = orientation === 'horizontal' && (isEnter(keyCode));
+			const isHorizontalArrow = orientation === 'horizontal' && changedBy === 'arrow' && (isRight(keyCode) || isLeft(keyCode));
 
-			if (isVertical || isHorizontal) {
+			if (isVertical || isHorizontal || isHorizontalArrow) {
 				this.pickerButtonPressed = 0;
 			}
 		}
@@ -712,7 +733,7 @@ const PickerBase = class extends ReactComponent {
 				Spotlight.setPointerMode(false);
 				Spotlight.focus(this.containerRef.querySelector(`.${componentCss.incrementer}`));
 			} else {
-				forward(`onSpotlight${cap(direction)}`, ev, this.props);
+				forwardCustom(`onSpotlight${cap(direction)}`)(ev, this.props);
 			}
 		}
 	};
@@ -740,7 +761,7 @@ const PickerBase = class extends ReactComponent {
 				Spotlight.setPointerMode(false);
 				Spotlight.focus(this.containerRef.querySelector(`.${componentCss.decrementer}`));
 			} else {
-				forward(`onSpotlight${cap(direction)}`, ev, this.props);
+				forwardCustom(`onSpotlight${cap(direction)}`)(ev, this.props);
 			}
 		}
 	};
@@ -759,11 +780,12 @@ const PickerBase = class extends ReactComponent {
 	};
 
 	determineClasses (css, decrementerDisabled, incrementerDisabled) {
-		const {className, joined, orientation, width} = this.props;
+		const {changedBy, className, joined, orientation, width} = this.props;
 		const {pressed} = this.state;
 
 		return classnames(
 			css.picker,
+			css[changedBy],
 			css[orientation],
 			css[width],
 			{
@@ -824,9 +846,13 @@ const PickerBase = class extends ReactComponent {
 	}
 
 	calcAriaLabel (valueText) {
-		const
-			{'aria-label': ariaLabel, joined, orientation} = this.props,
-			hint = orientation === 'horizontal' ? $L('press ok button to change the value') : $L('change a value with up down button');
+		const {'aria-label': ariaLabel, changedBy, joined, orientation} = this.props;
+		let hint;
+		if (orientation === 'horizontal') {
+			hint = changedBy === 'arrow' ? $L('change a value with left right button') : $L('press ok button to change the value');
+		} else {
+			hint = $L('change a value with up down button');
+		}
 
 		if (!joined || ariaLabel != null) {
 			return ariaLabel;
@@ -848,6 +874,7 @@ const PickerBase = class extends ReactComponent {
 		const {active} = this.state;
 		const {
 			'aria-valuetext': ariaValueText,
+			changedBy,
 			children,
 			css: incomingCss,
 			disabled,
@@ -897,7 +924,7 @@ const PickerBase = class extends ReactComponent {
 		const decrementIcon = selectDecIcon(this.props);
 
 		const horizontal = orientation === 'horizontal';
-		const isHorizontalJoined = horizontal && joined;
+		const isHorizontalJoinedEnter = horizontal && joined && changedBy === 'enter';
 
 		const reachedStart = this.hasReachedBound(step * -1);
 		const decrementerDisabled = disabled || reachedStart;
@@ -913,7 +940,7 @@ const PickerBase = class extends ReactComponent {
 			sizingPlaceholder = <div aria-hidden className={css.sizingPlaceholder}>{'0'.repeat(width)}</div>;
 		}
 
-		const showIndicators = horizontal && joined && Array.isArray(children) && children.length > 1;
+		const showIndicators = isHorizontalJoinedEnter && Array.isArray(children) && children.length > 1;
 		const valueText = ariaValueText != null ? ariaValueText : this.calcValueText();
 		const decrementerAriaControls = !incrementerDisabled ? id : null;
 		const incrementerAriaControls = !decrementerDisabled ? id : null;
@@ -923,6 +950,7 @@ const PickerBase = class extends ReactComponent {
 
 		if (joined) {
 			Component = SpottableDiv;
+			spottablePickerProps.changedBy = changedBy;
 			spottablePickerProps.onSpotlightDisappear = onSpotlightDisappear;
 			spottablePickerProps.pickerOrientation = orientation;
 			spottablePickerProps.spotlightDisabled = spotlightDisabled;
@@ -952,13 +980,12 @@ const PickerBase = class extends ReactComponent {
 				onKeyDown={this.handleKeyDown}
 				onKeyUp={this.handleKeyUp}
 				onUp={this.handleUp}
-				onMouseDown={this.handleMouseDown}
 				onMouseLeave={this.clearPressedState}
 				orientation={orientation}
 				ref={this.initContainerRef}
 				{...spottablePickerProps}
 			>
-				{isHorizontalJoined ?
+				{isHorizontalJoinedEnter ?
 					null :
 					<Cell
 						{...voiceProps}
@@ -1012,7 +1039,7 @@ const PickerBase = class extends ReactComponent {
 						</div>
 					)}
 				</Cell>
-				{isHorizontalJoined ?
+				{isHorizontalJoinedEnter ?
 					null :
 					<Cell
 						{...voiceProps}