+++ tests/ui/specs/TabLayout/VerticalTabsWithIcons/VerticalTabsWithIcons-specs.js
@@ -2,80 +2,77 @@ const Page = require('../TabLayoutPage');
 
 describe('TabLayout', function () {
 
-	beforeEach(function () {
-		Page.open('VerticalTabsWithIcons');
+	beforeEach(async function () {
+		await Page.open('VerticalTabsWithIcons');
 	});
 
 	describe('vertical tabs with icons', function () {
-
 		describe('collapsing/expanding behavior', function () {
 			describe('5-way interaction', function () {
-
-				it('should collapse tabs when focus is moved to a Spottable component in the content container - [QWT-2758]', function () {
+				it('should collapse tabs when focus is moved to a Spottable component in the content container - [QWT-2758]', async function () {
 					// 5-way down to second tab
-					Page.spotlightDown();
-					Page.tabLayout.view(2).waitForExist();
+					await Page.spotlightDown();
+					await (await Page.tabLayout.view(2)).waitForExist();
 					// Step 3-2: 5-way right to Spottable component
-					Page.waitTransitionEnd(1500, 'waiting for Panel transition', () => {
-						Page.spotlightRight();
+					await Page.waitTransitionEnd(1500, 'waiting for Panel transition', async () => {
+						await Page.spotlightRight();
 					});
 					// check that layout is collapsed
-					expect(Page.tabLayout.isCollapsed).to.be.true();
+					expect(await Page.tabLayout.isCollapsed).to.be.true();
 				});
 
-				it('should expand tabs when focus is moved to a Spottable component in the tabs container - [QWT-2758]', function () {
+				it('should expand tabs when focus is moved to a Spottable component in the tabs container - [QWT-2758]', async function () {
 					// 5-way down to second tab
-					Page.spotlightDown();
-					Page.tabLayout.view(2).waitForExist();
+					await Page.spotlightDown();
+					await (await Page.tabLayout.view(2)).waitForExist();
 					// focus the contents
-					Page.waitTransitionEnd(1500, 'waiting for Panel transition', () => {
-						Page.spotlightRight();
+					await Page.waitTransitionEnd(1500, 'waiting for Panel transition', async () => {
+						await Page.spotlightRight();
 					});
-					expect(Page.tabLayout.isCollapsed).to.be.true();
+					expect(await Page.tabLayout.isCollapsed).to.be.true();
 					// Step 4: Back to the tabs
-					Page.waitTransitionEnd(1500, 'waiting for Panel transition', () => {
-						Page.spotlightLeft();
+					await Page.waitTransitionEnd(1500, 'waiting for Panel transition', async () => {
+						await Page.spotlightLeft();
 					});
-					expect(Page.tabLayout.isCollapsed).to.be.false();
+					expect(await Page.tabLayout.isCollapsed).to.be.false();
 				});
 
-				it('should not show disabled tab contents when focused', function () {
+				it('should not show disabled tab contents when focused', async function () {
 					// 5-way down to second tab
-					Page.spotlightDown();
+					await Page.spotlightDown();
 					// 5-way down to third tab
-					Page.spotlightDown();
+					await Page.spotlightDown();
 
-					Page.tabLayout.view(3).waitForExist();
+					await (await Page.tabLayout.view(3)).waitForExist();
 
 					// 5-way down to fourth tab
-					Page.spotlightDown();
-					Page.delay(1000);
+					await Page.spotlightDown();
+					await Page.delay(1000);
 
-					expect(Page.tabLayout.view(3).isExisting()).to.be.true();
+					expect(await (await Page.tabLayout.view(3)).isExisting()).to.be.true();
 
 					// 5-way down to fifth tab
-					Page.spotlightDown();
-					Page.tabLayout.view(5).waitForExist();
+					await Page.spotlightDown();
+					await (await Page.tabLayout.view(5)).waitForExist();
 				});
 			});
 			describe('pointer interaction', function () {
-
-				it('should collapse and expand tabs when focus is moved between `Spottable` components in the content and tabs containers via pointer move - [QWT-2759]', function () {
+				it('should collapse and expand tabs when focus is moved between `Spottable` components in the content and tabs containers via pointer move - [QWT-2759]', async  function () {
 					// focus the layout's tabs
-					Page.tabLayout.hoverTabs();
+					await Page.tabLayout.hoverTabs();
 					// select the second tab
-					Page.tabLayout.tabItems[1].click();
-					Page.tabLayout.view(2).waitForExist();
+					await (await Page.tabLayout.tabItems())[1].click();
+					await (await Page.tabLayout.view(2)).waitForExist();
 
 					// move pointer to Spottable component in content container
-					$('#button2').moveTo();
+					await $('#button2').moveTo();
 					// check that layout is collapsed
-					expect(Page.tabLayout.isCollapsed).to.be.true();
+					expect(await Page.tabLayout.isCollapsed).to.be.true();
 
 					// go back to hover the tabs
-					Page.tabLayout.tabItems[0].moveTo();
+					await (await Page.tabLayout.tabItems())[0].moveTo();
 					// check that layout is not collapsed
-					expect(Page.tabLayout.isCollapsed).to.be.false();
+					expect(await Page.tabLayout.isCollapsed).to.be.false();
 				});
 			});
 		});