+++ tests/ui/specs/VirtualList/VirtualList/ElementSize/VirtualList-QWT-2609-specs.js
@@ -2,54 +2,54 @@ const Page = require('../VirtualListPage');
 const {expectFocusedItem} = require('../../VirtualList-utils');
 
 describe('spotlight size compare', function () {
-	beforeEach(function () {
-		Page.open();
+	beforeEach(async function () {
+		await Page.open();
 	});
 
-	it('should change spotlight size when item`s size changing [QWT-2609]', function () {
+	it('should change spotlight size when item`s size changing [QWT-2609]', async function () {
 		// Step 3 Verify: The default value for the 'itemSize' knob is itemSizeValue(default size is 156 for 4k) or half of 4k(78 for 2k).
-		const defaultItemSize = Page.getItemSize();
+		const defaultItemSize = await Page.getItemSize();
 		// The default size of Spotlight is 156 for 4k and 78 for FHD.
-		Page.buttonLeft.moveTo();
-		Page.spotlightRight();
-		const defaultSpotlightSize = Page.spotlightSize();
+		await Page.buttonLeft.moveTo();
+		await Page.spotlightRight();
+		const defaultSpotlightSize = await Page.spotlightSize();
 		expect(defaultSpotlightSize).to.equal(78);
 		// Step 4: Knobs > VirtualList > itemSize > 300
-		Page.inputfieldItemSize.moveTo();
-		Page.spotlightSelect();
-		Page.backSpace();
-		Page.backSpace();
-		Page.backSpace();
-		Page.numPad(3);
-		Page.numPad(0);
-		Page.numPad(0);
-		Page.backKey();
+		await Page.inputfieldItemSize.moveTo();
+		await Page.spotlightSelect();
+		await Page.backSpace();
+		await Page.backSpace();
+		await Page.backSpace();
+		await Page.numPad(3);
+		await Page.numPad(0);
+		await Page.numPad(0);
+		await Page.backKey();
 		// Verify item size
-		const curItemSize = Page.getItemSize();
+		const curItemSize = await Page.getItemSize();
 		expect(curItemSize.height).to.equal(150);
 		expect(curItemSize.width).to.equal(defaultItemSize.width);
-		Page.spotlightDown();
-		Page.item(2).moveTo();
-		expectFocusedItem(2);
-		const curSpotlightSize = Page.spotlightSize();
+		await Page.spotlightDown();
+		await (await Page.item(2)).moveTo();
+		await expectFocusedItem(2);
+		const curSpotlightSize = await Page.spotlightSize();
 		expect(curSpotlightSize).to.equal(150);
 		// Step 4: Knobs > VirtualList > itemSize > 50
-		Page.inputfieldItemSize.moveTo();
-		Page.spotlightSelect();
-		Page.backSpace();
-		Page.backSpace();
-		Page.backSpace();
-		Page.numPad(5);
-		Page.numPad(0);
-		Page.backKey();
+		await Page.inputfieldItemSize.moveTo();
+		await Page.spotlightSelect();
+		await Page.backSpace();
+		await Page.backSpace();
+		await Page.backSpace();
+		await Page.numPad(5);
+		await Page.numPad(0);
+		await Page.backKey();
 		// Verify item size
-		const newItemSize = Page.getItemSize();
+		const newItemSize = await Page.getItemSize();
 		expect(newItemSize.height).to.equal(25);
 		expect(newItemSize.width).to.equal(defaultItemSize.width);
-		Page.spotlightDown();
-		Page.item(4).moveTo();
-		expectFocusedItem(4);
-		const newSpotlightSize = Page.spotlightSize();
+		await Page.spotlightDown();
+		await (await Page.item(4)).moveTo();
+		await expectFocusedItem(4);
+		const newSpotlightSize = await Page.spotlightSize();
 		expect(newSpotlightSize).to.equal(25);
 	});
 });